Fix syncing bug related to integer overflow in syncrangesquared

+++ Data/Scripts/DetectionEquipment/Shared/GlobalData.cs
@@ -20,7 +20,7 @@ namespace DetectionEquipment.Shared
         public const ushort ClientNetworkId = 15287;
         public static int MainThreadId;
         public static double SyncRange => MyAPIGateway.Session.SessionSettings.SyncDistance;
-        public static double SyncRangeSq => MyAPIGateway.Session.SessionSettings.SyncDistance * MyAPIGateway.Session.SessionSettings.SyncDistance;
+        public static double SyncRangeSq => (double) MyAPIGateway.Session.SessionSettings.SyncDistance * MyAPIGateway.Session.SessionSettings.SyncDistance;
         public static readonly Guid SettingsGuid = new Guid("b4e33a2c-0406-4aea-bf0a-d1ad04266a14");
         public static readonly List<IMyPlayer> Players = new List<IMyPlayer>();
         public static IMyModContext ModContext;
